SCP-3343 fixed intermittent rollback failures in PAB
The cause of the rollback failure was that `Plutus.ChainIndex.UtxoState.rollbackWith`
computes the current tip of the index as the slot of the last transaction, ignoring
the fact that there may be empty blocks after that. When a rollback with a target
falls within such a sequence of empty blocks, the rollbackWith function
misinterprets this as a TipMismatch.

This change simply checks to see if the rollback is unnecessary because it is
through a sequence of empty blocks ending at the former tip.